backend/requirements.txt
alembic==1.13.3
amqp==5.3.1
APScheduler==3.10.4
async-timeout==5.0.1
bcrypt==4.2.0
billiard==4.2.1
blinker==1.8.2
celery==5.4.0
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
flask-core==2.9.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
loadenv==0.1.1
Mako==1.3.5
MarkupSafe==2.1.5
prompt_toolkit==3.0.48
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.0
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
zipp==3.20.2
